fix some cypress tests
