+++ integration_tests/e2e/recat/openConditions.cy.ts
@@ -105,9 +105,7 @@ describe('Open Conditions', () => {
     })
     cy.task('stubGetEscapeProfile', {
       offenderNo: 'B2345YZ',
-      category: 'C',
-      onEscapeList: true,
-      activeOnEscapeList: true,
+      alertCode: 'XEL',
     })
     cy.task('stubGetViolenceProfile', {
       offenderNo: 'B2345YZ',
@@ -352,42 +350,6 @@ describe('Open Conditions', () => {
         sequence_no: 1,
         risk_profile: {
           socProfile: { nomsId: 'B2345YZ', riskType: 'SOC', transferToSecurity: false, provisionalCategorisation: 'C' },
-          escapeProfile: {
-            nomsId: 'B2345YZ',
-            riskType: 'ESCAPE',
-            activeEscapeList: true,
-            activeEscapeRisk: true,
-            escapeListAlerts: [
-              {
-                active: true,
-                comment: 'First xel comment',
-                expired: false,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-14',
-                alertCodeDescription: 'Escape List',
-              },
-              {
-                active: false,
-                comment:
-                  '\nSecond xel comment with lengthy text comment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\n',
-                expired: true,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-15',
-                alertCodeDescription: 'Escape List',
-              },
-            ],
-            escapeRiskAlerts: [
-              {
-                active: true,
-                comment: 'First xer comment',
-                expired: false,
-                alertCode: 'XER',
-                dateCreated: '2016-09-16',
-                alertCodeDescription: 'Escape Risk',
-              },
-            ],
-            provisionalCategorisation: 'C',
-          },
           violenceProfile: {
             nomsId: 'B2345YZ',
             riskType: 'VIOLENCE',
@@ -496,7 +458,10 @@ describe('Open Conditions', () => {
       { question: 'Reasons that justify moving to open conditions?', expectedAnswer: 'Not applicable' },
     ])
     reviewRecatPage.validateVictimContactSchemeSummary([
-      { question: 'Does this prisoner have any victims opted in to the Victim Contact Scheme (VCS)?', expectedAnswer: 'No' },
+      {
+        question: 'Does this prisoner have any victims opted in to the Victim Contact Scheme (VCS)?',
+        expectedAnswer: 'No',
+      },
     ])
     reviewRecatPage.validateForeignNationalSummary([
       { question: 'Are they a foreign national?', expectedAnswer: 'No' },
@@ -558,42 +523,6 @@ describe('Open Conditions', () => {
             nomsId: 'B2345YZ',
             riskType: 'SOC',
             transferToSecurity: false,
-            provisionalCategorisation: 'C',
-          },
-          escapeProfile: {
-            nomsId: 'B2345YZ',
-            riskType: 'ESCAPE',
-            activeEscapeList: true,
-            activeEscapeRisk: true,
-            escapeListAlerts: [
-              {
-                active: true,
-                comment: 'First xel comment',
-                expired: false,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-14',
-                alertCodeDescription: 'Escape List',
-              },
-              {
-                active: false,
-                comment:
-                  '\nSecond xel comment with lengthy text comment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\n',
-                expired: true,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-15',
-                alertCodeDescription: 'Escape List',
-              },
-            ],
-            escapeRiskAlerts: [
-              {
-                active: true,
-                comment: 'First xer comment',
-                expired: false,
-                alertCode: 'XER',
-                dateCreated: '2016-09-16',
-                alertCodeDescription: 'Escape Risk',
-              },
-            ],
             provisionalCategorisation: 'C',
           },
           violenceProfile: {
@@ -831,7 +760,10 @@ describe('Open Conditions', () => {
       { question: 'Reasons that justify moving to open conditions?', expectedAnswer: 'Not applicable' },
     ])
     reviewRecatPage.validateVictimContactSchemeSummary([
-      { question: 'Does this prisoner have any victims opted in to the Victim Contact Scheme (VCS)?', expectedAnswer: 'No' },
+      {
+        question: 'Does this prisoner have any victims opted in to the Victim Contact Scheme (VCS)?',
+        expectedAnswer: 'No',
+      },
     ])
     reviewRecatPage.validateForeignNationalSummary([
       { question: 'Are they a foreign national?', expectedAnswer: 'No' },
@@ -893,42 +825,6 @@ describe('Open Conditions', () => {
             nomsId: 'B2345YZ',
             riskType: 'SOC',
             transferToSecurity: false,
-            provisionalCategorisation: 'C',
-          },
-          escapeProfile: {
-            nomsId: 'B2345YZ',
-            riskType: 'ESCAPE',
-            activeEscapeList: true,
-            activeEscapeRisk: true,
-            escapeListAlerts: [
-              {
-                active: true,
-                comment: 'First xel comment',
-                expired: false,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-14',
-                alertCodeDescription: 'Escape List',
-              },
-              {
-                active: false,
-                comment:
-                  '\nSecond xel comment with lengthy text comment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\n',
-                expired: true,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-15',
-                alertCodeDescription: 'Escape List',
-              },
-            ],
-            escapeRiskAlerts: [
-              {
-                active: true,
-                comment: 'First xer comment',
-                expired: false,
-                alertCode: 'XER',
-                dateCreated: '2016-09-16',
-                alertCodeDescription: 'Escape Risk',
-              },
-            ],
             provisionalCategorisation: 'C',
           },
           violenceProfile: {
@@ -1185,42 +1081,6 @@ describe('Open Conditions', () => {
         sequence_no: 1,
         risk_profile: {
           socProfile: { nomsId: 'B2345YZ', riskType: 'SOC', transferToSecurity: false, provisionalCategorisation: 'C' },
-          escapeProfile: {
-            nomsId: 'B2345YZ',
-            riskType: 'ESCAPE',
-            activeEscapeList: true,
-            activeEscapeRisk: true,
-            escapeListAlerts: [
-              {
-                active: true,
-                comment: 'First xel comment',
-                expired: false,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-14',
-                alertCodeDescription: 'Escape List',
-              },
-              {
-                active: false,
-                comment:
-                  '\nSecond xel comment with lengthy text comment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\n',
-                expired: true,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-15',
-                alertCodeDescription: 'Escape List',
-              },
-            ],
-            escapeRiskAlerts: [
-              {
-                active: true,
-                comment: 'First xer comment',
-                expired: false,
-                alertCode: 'XER',
-                dateCreated: '2016-09-16',
-                alertCodeDescription: 'Escape Risk',
-              },
-            ],
-            provisionalCategorisation: 'C',
-          },
           violenceProfile: {
             nomsId: 'B2345YZ',
             riskType: 'VIOLENCE',
@@ -1558,42 +1418,6 @@ describe('Open Conditions', () => {
             },
           ],
           socProfile: { nomsId: 'B2345YZ', riskType: 'SOC', transferToSecurity: false, provisionalCategorisation: 'C' },
-          escapeProfile: {
-            nomsId: 'B2345YZ',
-            riskType: 'ESCAPE',
-            activeEscapeList: true,
-            activeEscapeRisk: true,
-            escapeListAlerts: [
-              {
-                active: true,
-                comment: 'First xel comment',
-                expired: false,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-14',
-                alertCodeDescription: 'Escape List',
-              },
-              {
-                active: false,
-                comment:
-                  '\nSecond xel comment with lengthy text comment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\ncomment with lengthy text comment with lengthy text comment with lengthy text\n',
-                expired: true,
-                alertCode: 'XEL',
-                dateCreated: '2016-09-15',
-                alertCodeDescription: 'Escape List',
-              },
-            ],
-            escapeRiskAlerts: [
-              {
-                active: true,
-                comment: 'First xer comment',
-                expired: false,
-                alertCode: 'XER',
-                dateCreated: '2016-09-16',
-                alertCodeDescription: 'Escape Risk',
-              },
-            ],
-            provisionalCategorisation: 'C',
-          },
           violenceProfile: {
             nomsId: 'B2345YZ',
             riskType: 'VIOLENCE',
