+++ integration_tests/e2e/recat/womenEstate.cy.ts
@@ -66,9 +66,7 @@ describe("Women's estate recategorisation", () => {
     })
     cy.task('stubGetEscapeProfile', {
       offenderNo: testOffenderNumber,
-      category: 'T',
-      onEscapeList: false,
-      activeOnEscapeList: false,
+      alertCode: 'ABC',
     })
     cy.task('stubGetViolenceProfile', {
       offenderNo: testOffenderNumber,
