+++ integration_tests/mockApis/index.ts
@@ -7,6 +7,7 @@ import tokenVerification from './tokenVerification'
 import dpsComponents from './dpsComponents'
 import probationOffenderSearchApi from './probationOffenderSearchApi'
 import formApi from './formApi'
+import alertsApi from './alertsApi'
 
 export default {
   ...allocationManager,
@@ -18,4 +19,5 @@ export default {
   ...dpsComponents,
   ...probationOffenderSearchApi,
   ...formApi,
+  ...alertsApi,
 }
